moves keybinds to sep file, update readme
